# gpt/openapi.yaml
openapi: 3.1.0
info:
  title: SiteProof Audit API
  version: 0.2.0
  description: Evidence-based public website audit API for SiteProof GPT Actions.
servers:
  - url: https://YOUR-SITEPROOF-DOMAIN.example
security:
  - bearerAuth: []
paths:
  /api/discover:
    post:
      operationId: discoverSite
      summary: Discover robots, sitemaps and public URLs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: {type: string, format: uri}
                maxUrls: {type: integer, minimum: 1, maximum: 2000, default: 500}
      responses:
        '200': {description: Discovery result}
  /api/audit/page:
    post:
      operationId: auditPage
      summary: Audit one webpage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: {type: string, format: uri}
                mode: {type: string, enum: [fast, full], default: full}
      responses:
        '200': {description: Detailed page audit}
  /api/quickcheck:
    post:
      operationId: quickCheckSite
      summary: Run the five-point SiteProof QuickCheck
      description: Checks public DNS, website functionality, known subdomains, email DNS, and SSL/HTTPS. Private account and delivery facts are returned as access-dependent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickCheckRequest'
      responses:
        '200': {description: Five-area QuickCheck result and artifact-ready data}
  /api/audit/site/start:
    post:
      operationId: startSiteAudit
      summary: Start an asynchronous full-site audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteAuditRequest'
      responses:
        '202': {description: Audit job created}
  /api/jobs/{id}:
    get:
      operationId: getAuditJob
      summary: Poll a SiteProof audit job
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        '200': {description: Job state and compact audit result when complete}
  /api/audit/site:
    post:
      operationId: auditSiteSync
      summary: Run a synchronous website audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SiteAuditRequest'
                - type: object
                  properties:
                    responseMode: {type: string, enum: [compact, full], default: compact}
      responses:
        '200': {description: Audit result}
  /api/audits/{id}:
    get:
      operationId: getAudit
      summary: Retrieve a saved audit
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
        - in: query
          name: responseMode
          schema: {type: string, enum: [compact, full], default: compact}
      responses:
        '200': {description: Saved audit}
  /api/audits/compare:
    post:
      operationId: compareAudits
      summary: Compare two saved audits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [beforeAuditId, afterAuditId]
              properties:
                beforeAuditId: {type: string}
                afterAuditId: {type: string}
      responses:
        '200': {description: Resolved, introduced and persistent findings}
  /api/audits/{id}/export:
    post:
      operationId: requestAuditExport
      summary: Generate an audit export
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format: {type: string, enum: [json, csv, xlsx, docx, pdf]}
      responses:
        '200': {description: Export metadata and download path}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    QuickCheckRequest:
      type: object
      required: [url]
      properties:
        url: {type: string, format: uri}
        subdomains:
          type: array
          maxItems: 20
          items: {type: string}
          description: Optional known in-domain hosts or labels to verify.
        dkimSelectors:
          type: array
          maxItems: 10
          items: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,62}$'}
          description: Optional evidenced selectors; SiteProof does not guess selectors to claim DKIM absence.
    SiteAuditRequest:
      type: object
      required: [url]
      properties:
        url: {type: string, format: uri}
        mode: {type: string, enum: [fast, full], default: full}
        maxPages: {type: integer, minimum: 1, maximum: 150, default: 50}
        maxLinks: {type: integer, minimum: 1, maximum: 750, default: 750}
        respectRobots: {type: boolean, default: true}
        screenshotMode: {type: string, enum: [off, findings, all], default: findings}
        performancePages: {type: integer, minimum: 0, maximum: 10, default: 1}
        visualSamplePages: {type: integer, minimum: 0, maximum: 20, default: 5}
        includeExternalLinkChecks: {type: boolean, default: true}
        persist: {type: boolean, default: true}
        clientId: {type: string}
        projectId: {type: string}